# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This API Handle MELI - Sprint 1 - Grupo 2",
    "title": "MELI Bootcamp API - Sprint 1 - Grupo 2",
    "termsOfService": "https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones",
    "contact": {
      "name": "API Support",
      "url": "https://developers.mercadolibre.com.ar/support"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "description": "getAll products",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "List products",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "post": {
        "description": "Create ProductHandler",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "Create ProductHandler",
        "parameters": [
          {
            "description": "ProductHandler to Create",
            "name": "ProductHandler",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handlers.RequestCreateProduct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "description": "Get the details of a Products",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "Get ProductHandler",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProductHandler to be searched",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "delete": {
        "description": "Delete ProductHandler",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "Delete ProductHandler",
        "parameters": [
          {
            "type": "string",
            "description": "ID of a ProductHandler to be excluded",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "patch": {
        "description": "Update the details of a ProductHandler",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "Update ProductHandler",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Products to be updated",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated ProductHandler details",
            "name": "Products",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handlers.RequestUpdateProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    },
    "/api/v1/sections": {
      "get": {
        "description": "getAll sections",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sections"
        ],
        "summary": "List sections",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "post": {
        "description": "Create section",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sections"
        ],
        "summary": "Create SectionHandler",
        "parameters": [
          {
            "description": "SectionHandler to Create",
            "name": "SectionHandler",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.CreateSectionRequestDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    },
    "/api/v1/sections/{id}": {
      "get": {
        "description": "Get the details of a SectionHandler",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sections"
        ],
        "summary": "Get SectionHandler",
        "parameters": [
          {
            "type": "string",
            "description": "ID of SectionHandler to be searched",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "delete": {
        "description": "Delete SectionHandler",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sections"
        ],
        "summary": "Delete SectionHandler",
        "parameters": [
          {
            "type": "string",
            "description": "ID of a SectionHandler to be excluded",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "patch": {
        "description": "Update the details of a SectionHandler",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Sections"
        ],
        "summary": "Update SectionHandler",
        "parameters": [
          {
            "type": "string",
            "description": "ID of SectionHandler to be updated",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated SectionHandler details",
            "name": "Sections",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.UpdateSectionRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dtos.CreateSectionRequestDTO": {
      "type": "object",
      "properties": {
        "current_capacity": {
          "type": "integer"
        },
        "current_temperature": {
          "type": "integer"
        },
        "maximum_capacity": {
          "type": "integer"
        },
        "minimum_capacity": {
          "type": "integer"
        },
        "minimum_temperature": {
          "type": "integer"
        },
        "product_type_id": {
          "type": "integer"
        },
        "section_number": {
          "type": "integer"
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
...
    "dtos.UpdateSectionRequestDTO": {
      "type": "object",
      "properties": {
        "current_capacity": {
          "type": "integer"
        },
        "current_temperature": {
          "type": "integer"
        },
        "maximum_capacity": {
          "type": "integer"
        },
        "minimum_capacity": {
          "type": "integer"
        },
        "minimum_temperature": {
          "type": "integer"
        },
        "product_type_id": {
          "type": "integer"
        },
        "section_number": {
          "type": "integer"
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
...
    "handlers.RequestCreateProduct": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "expiration_rate": {
          "type": "integer"
        },
        "freezing_rate": {
          "type": "integer"
        },
        "height": {
          "type": "number"
        },
        "length": {
          "type": "number"
        },
        "netweight": {
          "type": "number"
        },
        "product_code": {
          "type": "string"
        },
        "product_type_id": {
          "type": "integer"
        },
        "recommended_freezing_temperature": {
          "type": "number"
        },
        "seller_id": {
          "type": "integer"
        },
        "width": {
          "type": "number"
        }
      }
    },
    "handlers.RequestUpdateProduct": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "expiration_rate": {
          "type": "integer"
        },
        "freezing_rate": {
          "type": "integer"
        },
        "height": {
          "type": "number"
        },
        "length": {
          "type": "number"
        },
        "netweight": {
          "type": "number"
        },
        "product_code": {
          "type": "string"
        },
        "product_type_id": {
          "type": "integer"
        },
        "recommended_freezing_temperature": {
          "type": "number"
        },
        "seller_id": {
          "type": "integer"
        },
        "width": {
          "type": "number"
        }
      }
    },
...
    "web_utils.response": {
      "type": "object",
      "properties": {
        "data": {}
      }
    }
  }
}